process location city and subdivision
Addresses issue #112. SQL Location table now stores city and subdivision as well
as country. When adding a location to the table, pipeline.py/interpreter.py use
location name to find the country code as before, but now also save the location
name in the "city" or "subdivision" field as well, as appropriate.